recompiled with new yacc ng go1.18

+++ flowd/parser.go
@@ -1219,7 +1219,7 @@ func (l *yyLexState) put_sqlstate(code string) bool {
 }
 
 //line yacctab:1
-var yyExca = [...]int8{
+var yyExca = [...]int{
 	-1, 1,
 	1, -1,
 	-2, 0,
@@ -1229,7 +1229,7 @@ const yyPrivate = 57344
 
 const yyLast = 274
 
-var yyAct = [...]uint8{
+var yyAct = [...]int{
 	212, 198, 146, 32, 123, 145, 139, 134, 147, 228,
 	59, 147, 225, 178, 227, 141, 140, 226, 136, 217,
 	65, 40, 216, 206, 215, 66, 214, 204, 54, 62,
@@ -1260,7 +1260,7 @@ var yyAct = [...]uint8{
 	33, 36, 8, 7,
 }
 
-var yyPact = [...]int16{
+var yyPact = [...]int{
 	73, -1000, 73, -1000, -1000, 180, -1000, 12, 9, 69,
 	178, 131, -1000, -40, -42, 176, -1000, 4, -1000, 4,
 	174, 169, 166, -1000, -1000, -1000, -1000, -1000, 147, 140,
@@ -1287,7 +1287,7 @@ var yyPact = [...]int16{
 	-1000,
 }
 
-var yyPgo = [...]int16{
+var yyPgo = [...]int{
 	0, 4, 67, 273, 272, 271, 270, 138, 226, 181,
 	269, 241, 268, 3, 267, 266, 265, 264, 263, 1,
 	262, 188, 261, 259, 258, 257, 7, 256, 255, 254,
@@ -1295,7 +1295,7 @@ var yyPgo = [...]int16{
 	243, 242, 237, 236, 0,
 }
 
-var yyR1 = [...]int8{
+var yyR1 = [...]int{
 	0, 20, 17, 17, 17, 17, 18, 18, 18, 18,
 	18, 21, 21, 21, 22, 22, 19, 19, 7, 7,
 	7, 7, 8, 8, 8, 8, 14, 5, 6, 6,
@@ -1310,7 +1310,7 @@ var yyR1 = [...]int8{
 	10,
 }
 
-var yyR2 = [...]int8{
+var yyR2 = [...]int{
 	0, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 	1, 3, 3, 3, 2, 3, 1, 3, 1, 1,
 	1, 1, 1, 1, 1, 1, 2, 3, 3, 3,
@@ -1325,7 +1325,7 @@ var yyR2 = [...]int8{
 	2,
 }
 
-var yyChk = [...]int16{
+var yyChk = [...]int{
 	-1000, -20, -10, -11, 9, 73, 16, -3, -4, 64,
 	11, 91, -11, -38, 39, -39, 94, 82, 94, 82,
 	18, 91, 85, 39, 17, 39, 69, 68, 99, 99,
@@ -1352,7 +1352,7 @@ var yyChk = [...]int16{
 	78,
 }
 
-var yyDef = [...]int8{
+var yyDef = [...]int{
 	0, -2, 1, 109, 89, 0, 92, 0, 0, 0,
 	0, 0, 110, 0, 0, 0, 95, 0, 97, 0,
 	0, 0, 0, 55, 56, 58, 59, 61, 0, 0,
@@ -1379,7 +1379,7 @@ var yyDef = [...]int8{
 	69,
 }
 
-var yyTok1 = [...]int8{
+var yyTok1 = [...]int{
 	1, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
@@ -1395,7 +1395,7 @@ var yyTok1 = [...]int8{
 	3, 3, 3, 99, 3, 100,
 }
 
-var yyTok2 = [...]int8{
+var yyTok2 = [...]int{
 	2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
@@ -1408,7 +1408,7 @@ var yyTok2 = [...]int8{
 	92,
 }
 
-var yyTok3 = [...]int8{
+var yyTok3 = [...]int{
 	0,
 }
 
@@ -1490,9 +1490,9 @@ func yyErrorMessage(state, lookAhead int) string {
 	expected := make([]int, 0, 4)
 
 	// Look for shiftable tokens.
-	base := int(yyPact[state])
+	base := yyPact[state]
 	for tok := TOKSTART; tok-1 < len(yyToknames); tok++ {
-		if n := base + tok; n >= 0 && n < yyLast && int(yyChk[int(yyAct[n])]) == tok {
+		if n := base + tok; n >= 0 && n < yyLast && yyChk[yyAct[n]] == tok {
 			if len(expected) == cap(expected) {
 				return res
 			}
@@ -1502,13 +1502,13 @@ func yyErrorMessage(state, lookAhead int) string {
 
 	if yyDef[state] == -2 {
 		i := 0
-		for yyExca[i] != -1 || int(yyExca[i+1]) != state {
+		for yyExca[i] != -1 || yyExca[i+1] != state {
 			i += 2
 		}
 
 		// Look for tokens that we accept or reduce.
 		for i += 2; yyExca[i] >= 0; i += 2 {
-			tok := int(yyExca[i])
+			tok := yyExca[i]
 			if tok < TOKSTART || yyExca[i+1] == 0 {
 				continue
 			}
@@ -1539,30 +1539,30 @@ func yylex1(lex yyLexer, lval *yySymType) (char, token int) {
 	token = 0
 	char = lex.Lex(lval)
 	if char <= 0 {
-		token = int(yyTok1[0])
+		token = yyTok1[0]
 		goto out
 	}
 	if char < len(yyTok1) {
-		token = int(yyTok1[char])
+		token = yyTok1[char]
 		goto out
 	}
 	if char >= yyPrivate {
 		if char < yyPrivate+len(yyTok2) {
-			token = int(yyTok2[char-yyPrivate])
+			token = yyTok2[char-yyPrivate]
 			goto out
 		}
 	}
 	for i := 0; i < len(yyTok3); i += 2 {
-		token = int(yyTok3[i+0])
+		token = yyTok3[i+0]
 		if token == char {
-			token = int(yyTok3[i+1])
+			token = yyTok3[i+1]
 			goto out
 		}
 	}
 
 out:
 	if token == 0 {
-		token = int(yyTok2[1]) /* unknown char */
+		token = yyTok2[1] /* unknown char */
 	}
 	if yyDebug >= 3 {
 		__yyfmt__.Printf("lex %s(%d)\n", yyTokname(token), uint(char))
@@ -1617,7 +1617,7 @@ yystack:
 	yyS[yyp].yys = yystate
 
 yynewstate:
-	yyn = int(yyPact[yystate])
+	yyn = yyPact[yystate]
 	if yyn <= yyFlag {
 		goto yydefault /* simple state */
 	}
@@ -1628,8 +1628,8 @@ yynewstate:
 	if yyn < 0 || yyn >= yyLast {
 		goto yydefault
 	}
-	yyn = int(yyAct[yyn])
-	if int(yyChk[yyn]) == yytoken { /* valid shift */
+	yyn = yyAct[yyn]
+	if yyChk[yyn] == yytoken { /* valid shift */
 		yyrcvr.char = -1
 		yytoken = -1
 		yyVAL = yyrcvr.lval
@@ -1642,7 +1642,7 @@ yynewstate:
 
 yydefault:
 	/* default state action */
-	yyn = int(yyDef[yystate])
+	yyn = yyDef[yystate]
 	if yyn == -2 {
 		if yyrcvr.char < 0 {
 			yyrcvr.char, yytoken = yylex1(yylex, &yyrcvr.lval)
@@ -1651,18 +1651,18 @@ yydefault:
 		/* look through exception table */
 		xi := 0
 		for {
-			if yyExca[xi+0] == -1 && int(yyExca[xi+1]) == yystate {
+			if yyExca[xi+0] == -1 && yyExca[xi+1] == yystate {
 				break
 			}
 			xi += 2
 		}
 		for xi += 2; ; xi += 2 {
-			yyn = int(yyExca[xi+0])
+			yyn = yyExca[xi+0]
 			if yyn < 0 || yyn == yytoken {
 				break
 			}
 		}
-		yyn = int(yyExca[xi+1])
+		yyn = yyExca[xi+1]
 		if yyn < 0 {
 			goto ret0
 		}
@@ -1684,10 +1684,10 @@ yydefault:
 
 			/* find a state where "error" is a legal shift action */
 			for yyp >= 0 {
-				yyn = int(yyPact[yyS[yyp].yys]) + yyErrCode
+				yyn = yyPact[yyS[yyp].yys] + yyErrCode
 				if yyn >= 0 && yyn < yyLast {
-					yystate = int(yyAct[yyn]) /* simulate a shift of "error" */
-					if int(yyChk[yystate]) == yyErrCode {
+					yystate = yyAct[yyn] /* simulate a shift of "error" */
+					if yyChk[yystate] == yyErrCode {
 						goto yystack
 					}
 				}
@@ -1723,7 +1723,7 @@ yydefault:
 	yypt := yyp
 	_ = yypt // guard against "declared and not used"
 
-	yyp -= int(yyR2[yyn])
+	yyp -= yyR2[yyn]
 	// yyp is now the index of $0. Perform the default action. Iff the
 	// reduced production is ε, $1 is possibly out of range.
 	if yyp+1 >= len(yyS) {
@@ -1734,16 +1734,16 @@ yydefault:
 	yyVAL = yyS[yyp+1]
 
 	/* consult goto table to find next state */
-	yyn = int(yyR1[yyn])
-	yyg := int(yyPgo[yyn])
+	yyn = yyR1[yyn]
+	yyg := yyPgo[yyn]
 	yyj := yyg + yyS[yyp].yys + 1
 
 	if yyj >= yyLast {
-		yystate = int(yyAct[yyg])
+		yystate = yyAct[yyg]
 	} else {
-		yystate = int(yyAct[yyj])
-		if int(yyChk[yystate]) != -yyn {
-			yystate = int(yyAct[yyg])
+		yystate = yyAct[yyj]
+		if yyChk[yystate] != -yyn {
+			yystate = yyAct[yyg]
 		}
 	}
 	// dummy call; replaced with literal code
